requirements.txt
Flask
flask-sqlalchemy
flask-login
flask-restful
flask-mail
PyJWT
marshmallow
marshmallow-jsonapi